requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
# passlib[bcrypt]==1.7.4
passlib
python-multipart==0.0.6
redis
celery
PyJWT==2.8.0
email-validator==2.1.1
africastalking==1.2.9
cloudinary==1.40.0
python-dotenv>=1.0.0